+++ examples/nextjs/.eslintrc.json
