+++ examples/nextjs/next-auth.d.ts
