+++ examples/nextjs/next-env.d.ts
