+++ examples/nextjs/next.config.js
