+++ examples/nextjs/package.json
