+++ examples/nextjs/public/favicon.png
