+++ examples/nextjs/src/components/layout.tsx
