+++ examples/nextjs/src/graphql/codegen.example.yml
