+++ examples/nextjs/src/graphql/document.graphql
