+++ examples/nextjs/src/graphql/schema/index.ts
