+++ examples/nextjs/src/graphql/urql.ts
