+++ examples/nextjs/src/pages/_app.tsx
