+++ examples/nextjs/src/pages/api/auth/[...nextauth].ts
