+++ examples/nextjs/src/pages/api/graphql.ts
