+++ examples/nextjs/src/styles/globals.css
