+++ examples/nextjs/src/utils/server-side-fetch.ts
