+++ examples/nextjs/tailwind.config.js
