+++ examples/nextjs/tsconfig.json
