--- examples/todo-nextjs/grafbase/schema.graphql
@@ -1,12 +1,0 @@
-type TodoList @model {
-  id: ID!
-  title: String!
-  todos: [Todo]
-}
-
-type Todo @model {
-  id: ID!
-  title: String!
-  complete: Boolean
-  list: TodoList
-}
